requirements.txt
PyYAML==3.11
boto3==1.3.1
pep8==1.7.0
pylint==1.6.4
botocore==1.4.41
backports.functools_lru_cache==1.2.1
